requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
yarl==1.11.1